Workaround to make theme editor tooltips lazy.
Unfortunately there is no support for lazy tooltips in Swing so this
patch will delay the loading as much as possible. To decide if a tooltip
must be loaded, waits until the mouse is over the table and checks if
it's over the current cell.
Also it adds caching for tooltip loading to avoid reloading the same
tooltip over an over when switching between cells.

Change-Id: I5129d83231277694e3e8a6579e7cc5132ec0a149