Don't launch activities for watch face applications.
When an application is created, a run configuration is created for it.
This run config defaults to installing the apk and then launching the
default launcher activity of the application. Watch faces however don't
have a default activity (they are similar to a wallpaper service). So
the default run config shows an error saying that it can't find any
activities.

This CL attempts to detect whether an app is implementing a watch face
by examining its manifest, and if it detects a watch face app, then
the default run config is set to not launch any activities - it'll
just install the apk.

In the future, the framework will provide a way to "launch" watch faces,
possibly via the watch face preview mechanism. When this support becomes
available, the launch action should be changed from just installing
the apk to installing and then launching the preview activity.

Fix for https://code.google.com/p/android/issues/detail?id=151353

Change-Id: Ia4522830da9320e6c30a7b9ff3d63579e602a99c