navigator: Remove custom tree builder
We were using a custom tree builder to handle the fact that a single
node could correspond to multiple PsiFiles or PsiDirectories. However,
it turns out that this is handled by the existing tree builders as
long as the contains() method is implemented correctly on each of
those nodes. So the custom tree builder was actually unnecessary.

Change-Id: I7d378a8f3208240da9f862f50880af1f9044aa66