Adds experimental CPU monitoring to Android Tool Window.
By executing 'cat /proc/[pid]/stat' and 'cat /proc/stat' in an ADB
shell, this new CPU usage monitor parses the output and grabs the user
and kernel CPU schedule information and computes the % CPU usage.
It then displays this information in a dedicated CPU tab in the
Android Tool Window, similar to the memory monitor.

This change also includes some refactoring of the memory monitor such
that both the cpu and memory monitor can share code.

Change-Id: I54f711cfe7b9379465222b054c181d1b4a93176c