Move model cloning and commiting to data service classes.
As the mode cloning and commiting are heavy weight operations,
instead of getting modifiable model and commiting the changes to it
in each customizer, all the model cloning and comming operations
are moved to data service classes to perform a single clone and
commit operation per data service.

This change reduced the time to execute all android project
customizers with AndroidProjectDataService from 37,352 ms to 2,824 ms
during a sync operation on a project with more than 300 modules.

See https://code.google.com/p/android/issues/detail?id=123783

Change-Id: Iba0a3c835108d9bbd038bb41a917948c9c5eb902