Add checks to in-memory R class.
Because the in-memory R file is not on disk, it does
not get enumerated when searching files for find usages.
This change adds the psiclass representing the in-memory R
to the facet and also to ResourceUtil so it can be found.

There is also some refactoring of findResourceFieldsFromClass
so that there is less code duplication.

If the in-memory R class is never set (it currently is *not* set
in the android plugin), this change will result in no net
difference in behavior.

Change-Id: Iebb9c181418afc86f31daa0b188591a02dc72c4e