Another fix for dynamic wizard size change
https://android-review.googlesource.com/137871 changed the way step
components are added to wizards, and
https://android-review.googlesource.com/138541 added support for a
further case. However, in some cases (e.g. the first run wizard)
the wizard is never "shown" at all; rather the wizard component is
added directly to another container (more or less necessary in that
case, since the window and content are created by intellij code).

This change puts back the original behavior, such that components
will be added to the wizard when the step is shown, if they haven't
already been added.

Change-Id: If4a2278f6f5ad6f6a85c7892cc0c47e46f299a9b