Enable "Show bitmap" action that went missing in 1.2
Studio 1.1 had a feature to right click on a bitmap in the debugger
variables pane and view the actual bitmap on the device. IJ 14 has
a revamped debugger implementation, and this feature possibly stopped
working after that rewrite.

This CL re-enables it. Rather than using the custom action as done
earlier, we now follow the same approach as used in IJ's "View icon"
action that shows up for Java images.

Fixes https://code.google.com/p/android/issues/detail?id=77414 and
https://code.google.com/p/android/issues/detail?id=159479

Change-Id: Ifce6df615eb34d7dbdc0717327344757bc9b57c5