Minor, trivial cleanup.
Instead of using a Map<String, List<String>>, I changed it
to Multimap<String, String>.
Also, instead of keeping an empty map in the Project, I set
it to null.

These are tiny changes to code that is working perfectly
well. I made them since I'm about to touch this area while
fixing a bug related to dependency setup.

Change-Id: I60c30e82aba9b18b953eab1d83be0fd7aba11ed8