Removed duplicated "dependency setup" errors on sync.
The issue was that if a module was missing during project
sync, and there were more than 1 module depending on the
missing module, the IDE showed duplicated errors, and each
error did not say which modules were the affected.

The problem in the code is that we were displaying errors
in the IDE as soon as we found them.

The fix involves storing all the error messages, removing
duplicates and storing the names of the dependents. After
sync is finished, we show the errors.

The issue below has before/after screenshots.

Fixes
https://code.google.com/p/android/issues/detail?id=160017

Change-Id: I8d2f3df65aac634d31b7953249e6e92708586138