Fix resource repository updates
In 14.1 the PSI event delivery for XML document edits has changed
slightly, and one of the unit tests uncovered a bug here where
an edit which adds an id to a layout would not mark the resource
repository as having a new generation count. That's fixed by this
CL (and now the test passes).

The changed PSI events also caused another bug to fail; that was
a test bug (which relied on certain edits causing the repository
updater to switch from accurate updates to a delayed full file
rescan). The events are now more accurate, so the updater is able
to handle it incrementally, and the test didn't expect that. It now
formally forces the repository to go into pending scan mode, such
that it can correctly check that it can synchronously flush pending
changes.

Change-Id: I71ccb342bf03e153f2d891202118091656376ec0