Now the IDE attempts to find and include any missing modules.
When importing a subset of a project, the selected modules may
depend on other modules that were not included. In the case of
Android libraries that were not excluded, the IDE will link
modules in the project subset against the compiled version of
such missing libraries. Unfortunately, for Java libraries there
is no information during project setup about the compiled
version of the library (this is being tracked in Issue 123787).

The proper fix needs to happen in the Android Gradle plugin, but
until then, and for older versions of the plugin, the IDE can
find and include the modules that were previously excluded.

Fixes issue
https://code.google.com/p/android/issues/detail?id=160579

Change-Id: I239f09a1794e6026a1ef4e5cbc8e3dfffe747f07