Change gradle tasks executed when running unit tests.
This applies to the Gradle-aware make step. The only task that we
need to call is compileDebugUnitTestSources, this will trigger all
other necessary tasks on the gradle side. See the comment in
GradlInvoker on why we should not call the main compile task (in
short: because Jack).

This CL also makes sure that all relevant (as defined by the Build
Variants tool window) classes folders are up-to-date. This cannot be
done from Gradle, since gradle (almost) always uses the relase variant
and doesn't know anything about the AS workspace settings.

This fixes a problem where unit tests that use classes from a dependency
android library can't find those classes. I consider this a big win.

The solution is inconsistent with Gradle (AS uses debug classes when
running tests, gradle uses release), but consistent with the code
completion in AS. I consider this a minor problem, as compared to the
"big win" from previous paragraph.

To cover this functionality, I switched to a more complicated project
for UI testing. This is essentially the same as we use for gradle
integration tests. All expected functionality still works with both Make
steps.

Bug: http://b.android.com/156256
Bug: http://b.android.com/159814
Bug: http://b.android.com/172409
Change-Id: Ia68d866731a4295439dc1e567031c2fa1708ec50