Added error message for successful syncs with errors.
There are cases that project sync succeeds, with errors.
For example, when a dependency is missing, Gradle will
return the project structure and report an error.
The problem is that the IDE says "Gradle project sync
failed. Basic functionality (e.g. editing, debugging) will
not work properly." which is misleading, because editing
is possible because there is a project structure. The
error message needs to reflect that sync succeeded with
errors.

Fixes:
https://code.google.com/p/android/issues/detail?id=160903

Change-Id: Iacf0afc8120c56ad4692924beadbe0f55a634087