gradle: Fix broken test on Jenkins that depend on static content.
The test: AndroidGradleBuildProcessParametersProviderTest was failing
on Jenkins but succeeded locally. The problem was that other tests
were adding to the available JDKs and this test depend on getting the
JDK that is setup by the test. Fix this problem by removing all JDKs
before the test is run.
Also change several assertTrue to assertEquals such that a future
failure is easier identified.

Change-Id: Ic1484b05c3f3bca8865d5a37143e2af297a0662c
(cherry picked from commit a3df98a4858d986b21b39f58447ac8081e93964d)