Update resource type inspection to handle new annotations
This CL generalizes the resource type inspection (which already
handled not only resource types but @IntDef checks as well) to
look for and report some additional issues now flagged by
lint. It also removes the corresponding lint checks from the
issue registry. (This is done such that we don't have to look
up annotations on all calls both from this inspection as part
of resource type checks and then again from the lint check.)
The color-as-usage and check-permission checks are subsumed
under the new checks so they are also removed from the
inspections registry.

It also updates the resource type unit test to use a different
mechanism for checking the results. This allows us to inline
the source code, which makes it easier to create, validate and
update the tests.

Change-Id: Ieadd7e38626fa955e5267e746f558562e1234ff6