Reducing dependencies on sdklib.internal.repository Part 1
The existing com.android.sdklib.internal.repository is difficult
to work with, and a lot of the functionality is not necessary in
the Studio restart-and-install model. I hope to be able to provide
an alternate implementation that can be injected into the main sdk
code to better fit our needs.
To that end, I hope to refactor studio code such that it doesn't use
com.android.sdklib.internal.repository at all.
This is the first step of that change, dealing with the dependencies
in SystemImageList and AvdWizardConstants.

Change-Id: Ib4c971b8bb3b80099ef56d40abd5bffb043a90ff