Perform reproxy operation in a background thread.
I tried to replace the existing reflection based proxy util with the
real implementations of the model classes, but that reduced the time
only by 10%. That indicates that the problem is not with the proxy
code, but with massive size of the model object we are trying to
clone.

In this change, the proxy operation is moved to a background thread
and getting executed in parallel with the project sync operation,
so that it will not add any latency to the sync operation.

Fix for https://code.google.com/p/android/issues/detail?id=161147

Change-Id: I8ae643282673336d2f09e388f968b25a92e773cb