Removing studio references to code that was branched
After this change, RemoteSdk, RemotePkgInfo, Update, UpdateResult,
and UpdateTest (as well as some other code) in sdklib will no
longer be used, and can be deleted in a subsequent change.

After that the only code related to remotes in sdklib will be what is
used exclusively by the old sdk manager, at which point it will all be
marked as deprecated.

Change-Id: I98c3dde113de70d4ccdf56b5602051959e527cfe