Properties framework
Properties are, in a nutshell, values that can be queried, set, and
chained in a fluent manner.

Properties are designed to ultimately replace the data binding logic
provided by TemplateWizardState, ScopedStateStore, and
ScopedDataBinder from the wizard package. Those classes, while
powerful, are tied to the wizard code, whereas this new approach
should be useful in any context, UI or not, where you might benefit
from having rich data interactions.

This CL lays down the core groundwork which will be further fleshed
out in followup CLs.

Change-Id: I42be4ef0720620c0eca56201c04894e807ae87a1