Added initial Property expressions
An expression is an observable value that wraps (and modifies) some
target observable value. For example, "NotExpression" converts a
boolean value to its inverse value, while "IsEmpty" converts a
string to a boolean.

This CL provides both the groundwork for supporting future
expressions as well as provide examples for how to create more
expressions later.

Change-Id: I15931a39b3050c1f0472cd95a35711a064027dee