GUI tests now can be executed in IDEA 14.1.
This CL fixes ClassLoader issues that prevented GUI
tests from being executed from IDEA 14.1.

The problem was that the ClassLoader that has the URLs
of the plugin jars was loaded by another ClassLoader,
and IDEA's bootstrap code was ignoring it, resulting
in class loading exceptions when starting up Android
Studio in a GUI test.

The fix is to access the URLs of that ClassLoader using
reflection.

Change-Id: I58c71cac1b4b71ea27d3ab77e23ab8342747b302