Refresh VFS when reading manifest copy.
This fixes an issue where the temp copy of a provided custom compiler
manifest file (created through native File operations) was sometimes not
picked up by a VFS read operation immediately afterwards, despite the
copy being present and readable on disk. This resulted in failure to
launch run configurations with an error saying the manifest could not be
found.

Change-Id: I2604cf59a683256e2fd6050dc59d182e254f8c56