Fixed project sync when a Java lib module depends on another Java lib module.
The issue was that after CL
https://android-review.googlesource.com/146886
we create a proxy for IdeaJavaProject and it worked well
when the Java lib had library dependencies but we never
tested proxing when IdeaJavaProject has dependencies on
other Java modules.

When dealing with inter-module dependencies among Java
modules, creating a proxy of IdeaJavaProject fails because
ProxyUtil does not know how to handle DomainObjectSet, a
collection that comes from Gradle and it is being used in
IdeaModuleDependency (an instance of IdeaModuleDependency
is retrieved from Gradle on project sync.)

Instead of allowing proxying of DomainObjectSet, which we
really don't need because the information in those collections
is discarded, we create a new type of "module dependency" that
only has the information we need and it is serializable and
understood by ProxyUtil.

Fixes
https://code.google.com/p/android/issues/detail?id=167378

Change-Id: Ie6175b36b9719680d745de686f2ce50ec6f7c3c5