Fix exception thrown with custom arraymap renderer
Studio has a custom renderer for ArrayMaps. This renderer evaluates the
following expression for each entry in the ArrayMap:
 new Object[] {this.keyAt(%1$d), this.valueAt(%2$d)}
In Studio 1.2, this fails with an exception thrown by the runtime saying
that the Object[] class is not found. Until we have a complete fix for
this issue, this CL provides a temporary patch by not creating a new
Object, but just showing the value for each map entry (so we miss out on
displaying the key).

Partial fix for https://code.google.com/p/android/issues/detail?id=168873

Change-Id: I5c5e97197787fa497c6934687a2f46de3b937d80