[UiTests] Use JDK_HOME_FOR_TESTS to pin a specific JDK version.
It may be useful to be able to run the same suite against different JDKs. Also
right now we let Studio pick the JDK, which is not hermetic w.r.t. the state of
the local machine (I have JDK9 installed and all the tests fail early because
MaxPermSize doesn't exist anymore).

Change-Id: Id741f05d2c9dae17ff66ec4aa7929beed256beee