163131: Allow studio to come up when we have no internet.
Bug report: https://code.google.com/p/android/issues/detail?id=163131
The first start wizard caused Android studio to crash when an
invalid proxy was entered when no internet could be detected.
Now this exception is caught and the user is allowed to complete
the wizard. If compatible sdk was not found then the wizard will
attempt to run again next time studio is started.

Change-Id: I8e29c4bc80103c7b67984d7da1cea4ef30d4d7d5