Removes Heap tabs in HprofViewer
Heap tabs will be replaced by a filtering mechanism
that works at instance level, not class level.

Renamed heap table to be class table and made it show
all the classes.

Removed HeapTableManager as there is only one table needed.

Changed HprofViewPane to be the one that does all the linking.

Change-Id: Ib30addea618a10274331ac5701f7000d2494837b