Now messages from exceptions thrown when building project are not discarded.
The problem was that the exception thrown by the build was being caught
by the IDEA Gradle framework, and the real cause was being discarded.

The fix is to catch the exception before IDEA does, and obtain and display
the real error message. Now we also show the error in the "Messages"
view, instead of a popup dialog (improved user experience.)

Fixes
https://code.google.com/p/android/issues/detail?id=130520

Change-Id: I829e903dff82af80ebedc954eb7620ea2db0d3e7