Fix failing tests after updated test data.
Bitmap completion test failed because the new public.xml exposes
tileModeX and tileModeY attributes, while the test expected only
tileMode as available for completion.

an3_after.xml: Fails because the test expects "background" to
be the only completion option for "backg" prefix, but API 22
introduces attributes backgroundTint and backgroundTintMode.

fragmentCompletion3.xml: Fails because the test expects "ta" to
complete to "tag", but the new attributes
fragmentAllowEnterTransitionOverlap and
fragmentAllowReturnTransitionOverlap both have "ta" prefix in
them which causes it to fail.

Change-Id: I1d17fab8f0b7364a9be052d46573bbc96cf37827