Nele: Add speed search to the palette.
Add speed search and make the tree the receiver of the focus when
the palette window is chosen. The latter allows speed search to
be started after the user clicks on the titlebar of the palette.

Change-Id: Ifbf2a815ffcebf939fbe1a1fb0286bf834434a98
(cherry picked from commit 872bdcf1cca16cdd09b1998a2d56fa0a1116de4e)