Nele: Add speed search to the palette.
Add speed search and make the tree the receiver of the focus when
the palette window is chosen. The latter allows speed search to
be started after the user clicks on the titlebar of the palette.

Change-Id: Ifbf2a815ffcebf939fbe1a1fb0286bf834434a98
(cherry picked from commit 677e348c98e6d844f40344ae63f34f192cc6b278)