Developer services now restore old values if editing is cancelled
A developer service has an associated context which a user can modify
through the UI in the project structure dialog. However, a user may
press cancel on that dialog and expects all values to restore to
their previous value.

This CL adds that functionality, by updating the ServiceContext
class to have the ability to snapshot and restore values.

Change-Id: I1b3d80967d266061fd595d61d9d97fd5c596cb65