Fixing excessive refreshing of remote sites in sdk config
When config panels are initialized reset() is called to allow them
to fetch their content. Removing unnecessary fetch that was done
before that.
Fixed loadSynchronously to actually wait until an existing fetch
is complete before returning in the case we're not on the dispatch thread.
Fixed assignment to myPackages in SdkState such that we'll always get at
least empty packages, and during refresh (when run for the first time)
we'll get the local packages right away.
Changed the updater checker to not force a refresh of sources.
Added updating message to sites page.

Change-Id: Iec6d21aef56f8e0055edbf805db576a1e0cb187d