Add resource notification manager
The resource notification manager handles tracking changes that
can affect resource editors. We currently have many independent
implementations of this in various editors; this should isolate
it all into a single sharable piece (which also has the advantage
that it (for example) only installs a single PSI listener, and
only while editors are actually active. (The layout editor for
example adds a separate PSI listener for every single file,
and they stay in the PSI listener queue even when the file is
in the background.)

Updating existing editors will be done in separate CLs.

Change-Id: I4d67ad81b9497dbe3190b3bf2324e98b5ba1e7ae