Adds ColumnTree builder
A ColumnTree is a replacement for TreeTables that instead
of being a Table (that uses a tree to render a column), a ColumnTree
is a normal tree that uses a table header to drive its rendering.

The rendering of the table header and the tree is completely separate,
except for the fact that the tree renderer listens to column
changes of the empty table to re-render its components.

This allows standard tree utilities to be used on these components
(like SpeedSearch).

ColumnTree don't support all functionality from tables (yet), but
offer a much simpler interface and setup.

Change-Id: Ifef84e7710ee89bc510b69d6f23c17e3a82d6a54