More rendering fixes.
Some R inner classes (like <package>.R$attr) failed to load when
rendering preview for new projects. The reason was that the library
corresponding to the R class wasn't registered. The issue only occurs
when using gradle plugin version 1.2.x, as the location of classes.jar
has changed and the code wasn't updated to match it.

Also, the workaround for issue http://b.android.com/164378 wasn't shown
always. Fix it by adding the message when the bug manifests in the form
of broken classes.

Change-Id: I3977ef598c92a3853c30de87765b440dfaa0e36a