Now parsing recipe.xml using JAXB instead of SAX
Whereas SAX gives you raw element names and attributes, JAXB
deserializes XML into Java objects.

This lets us break recipe logic out into separate classes, the list
of which are returned as a Recipe which can be queried and/or
executed on demand. This also allows us to remove the dry-run
approach.

Change-Id: I7e6295311bd09e732a25e086247c24762d2ca756