Ensure that compileSdkVersion uses int, not targetname
The code was relying on IAndroidTarget#getOptionalLibraries
returning null when there are no optional libraries. The
sdklib implementation recently started returning an empty
list instead, which the template code would take to mean
"this has libraries" so it would use android-22 instead of 22
for example.

However, we should be calling isPlatform() rather than
relying on the libraries of the target.

Change-Id: I4fa1b758046a022cba3a7e90f00fff48bbcff6ca