Fixed broken NewProjectWizardTests
A recent refactoring to template.java broke the ability for us to
parse some "weird" filenames (particularly, paths with " and < in)
them. It turns out, our project creation flow doesn't allow for us
to create projects with those invalid characters in the path anyway,
so the safest course of action here is to disable the tests.

Change-Id: I8d5abdd1c4d998c8b98818bd2da606b1fa4e5eda