Adds expandability to long lists of objects in Instaces view.
Instances list can get very long (>10k items), and this causes a
noticable slowdown in instances tree creation. This CL adds a new node
that the user can press to load 100 instances at a time (starts with
100 instances displayed).

This CL also adds a couple of usability features/fixes, such as
retaining the selected class when the heap changes.

Change-Id: I869936698a2aa9950d0c6b0558d788105602f64b