Nele: Drop the new component on the correct screen view.
There were a bug: if you drag a new component in from the left
(i.e. over the blueprint) and drop it on the normal screen view,
the coordinates would be computed as if it was dropped on the
blueprint view.

Change-Id: Ib5254c534e7457541211db089a543a163e0fcc27
(cherry picked from commit a810fa1b6e4946908e2ae1a11341123a5f67ba9c)