Improved service editing flow
Service editing used to be a bit more fragile, with a "modified"
property that anyone could edit. But now we expose more restricted
beginEditing/cancelEditing/finishEditing methods. Plus there are
now callbacks in ServiceContext so services can initialize and
validate themselves, so you don't try and install an invalid
service.

Change-Id: I4c709f65d835277fb503e8c1ad6a0763f59cb537