Removed InitializeXmlParser
The feature that this class existed to satisfy, initializing a
service before showing it to the user, got cut from the 1.3
launch - now, we just show a service as installed or not, without
any details.

We'll put this back in for 1.4, but we need to think about this
feature a bit first, as parsing arbitrary projects for values is
not an easy problem.

Change-Id: Ib724733661dcbcd367da72db098f1edd814a55a5