Adding more expansion nodes and sorting to instances view.
This CL adds sorting to instances view. This is exceedingly convoluted
due to the dependency on the DebuggerTree, which results in writing
around the limitation of private fields employed by the debugger
infrastructure. Also, the sorting has to have numerous exceptions and
fallbacks, including:
* always sorting expansion nodes to the bottom, no matter the sort
    order
* fall back to memory ordering/instantiation-time ordering when sort
    result ends up being equivalent (e.g. for array elements, the
    ordering should be numeric, not lexicographical)
* resorting the list on node creation (this is a TreeBuilder, so not
    all nodes are present at all times)
* work with a tree structure, and not just a table structure.

In addition, this CL adds expansion nodes to array instances (some
arrays have a lot of elements).

Change-Id: Ib7e0fafe755f1161a5130d000b583944da7c2e20