Make LayoutPreviewTest pass
Updates the LayoutPreviewTest such that all but one test cases
pass (the last one is blocked by bug 174236 and is disabled for now).

Some of the rendering tests failed because they're using too-old versions
of the tools. This CL updates all the tests to use recent build tools,
compile SDKs and updated support library dependencies.

Change-Id: I42993aacaf0a320a763f18dabfb26817edca49c2