Added support for service.xml to use the <open> command
Recipe.xml normally supports the open command, which opens a target
file in the editor after the recipe executes. However, it doesn't do
it itself, but instead, passed the instruction on and hopes it is
handled externally. Services did not handle this request before but
they do after this CL.

Change-Id: I875b10e59093c77af2a57214583b55c6b15efa25