Add getter metadata and variable+call lookups for type computation
For fields without annotations, look up the corresponding
getter method (if any) and use its annotations. This is
useful since fields tend to be private (and therefore not
annotated; certainly not in framework code with external
annotations), whereas the getter (if found) will often
provide metadata about the field.

For local variables, search through variable declarations
and initializers (recursively, if necessary, to deal with
intermediate variables) and use annotations from methods
(in method calls), field assignments and variable
declaration types.

Change-Id: I3695002c6e168d680e9d0d03806e14fc71914c08