Create ThemeEditorContext to store Configuration and Module
For properly supporting themes that come from different
modules in the theme editor, we need to update Configuration
instances that are stored in quite a lot of places right now.
As a first step for doing that, this change introduces
ThemeEditorContext class which contains method to acquire
instances of Configuration, Module and other classes that
are used all over the theme editor.

Later, when theme editor would show all themes from the
project, to update Configuration one would need just to call
setConfiguration on the only instance of ThemeEditorContext
instead of propagating updated value to all its users.

Change-Id: Ib1e2406500795698f38c092c1732c5cc81963462