ClassConverter: make _Original methods private.
This prevents StackOverflowError when there are two custom views which
implement onMeasure (or any of the other modified methods), one inherits
from the other and the subclass calls super.onMeasure. Earlier, the call
to super's onMeasure_Original resulted in calling subclass's
onMeasure_Original, which called super.onMeasure() and hence the
StackOverflowError.

Also added tests and moved the initialization of error method
description to class initializer.

Change-Id: I323c9c13aed1fe881f458c50e3333610d4bbabfd