Preliminary support for whole-project Theme Editor
This CL makes theme resolution in theme editor traverse the
whole project (instead of single module which was the case
previously). Themes list now shows themes from all modules,
and after switching a theme it's rendered in context of its
source module.

Things that are planned for next CLs:
* Get rid of ThemeResolver class
* Fix the way themes are resolved
* Refactor ThemesListModel to get rid of instanceof checks
* Make ProjectThemeResolver to be a static method instead

All these issues marked with TODO comments in the code.

Change-Id: Ib8ea53d7fec1ee2e718ce28a3b9c51cb35dad58a