Preview packages incorrectly picked up by welcome wizard
Previously, ComponentInstaller (strangely, but that's a different
story...) creates it's own SdkPackages, and treats all packages
returned by getUpdatedPkgs() as valid updates to be installed.
This isn't entirely correct, as this includes both "regular" updates
and updates to preview versions. The effect in this case is that the
normal revision would be requested and installed, but then when
checking whether the install was successful we would always see the
preview as an available update, and thus assume that the install
failed.
Now we only include updates that are non-preview->non-preview
or preview->preview.
Change-Id: I0af928bdab1eb1da0b209db81d18bb47b0cd40e0
(cherry picked from commit df3357d77a1ad40b7ce66ac064e695efd02fca3f)