Potentially better error messages in StyleResolver.getStyle
StyleResolver.getStyle used to violate @NotNull contract by
returning null at the end of its body, which is going to
cause runtime exception during the execution (as @NotNull
contracts are enforced). This CL changes is so that this
function throws an exception explicitly, which would
more clearly indicate source of a problem in an event when
it occurs.

Change-Id: I578a8859d8622a2d7dce63afa8e7a1bb54527b7b