Logcat: Fixes filtering issues
Fixes several issues that would cause the active filter not to
be reflected in the combobox.

The combobox now contain the actual ConfiguredFilters and selecting
an item in the combobox is the only way to change a filter in the console
model.

The method that used to change the console model filter is now gone, and
all changes that want to be done to the current filter need to go through
the combobox.

Change-Id: I69f406a60525e825d3043e4830ecc7c744967527