Support computing error suggestions in a background thread
When you're typing in the XML editor while layout preview
window is showing, the layout is periodically updated. This happens
in a background thread without holding a read lock (since if a read
lock was held during rendering, typing a character would not allow
a write lock for the edit to be acquired, so there would be a
noticeable lag while typing.)

However, if there is a render error, the error handler opens up
an error suggestion window. Computing these suggestions is sometimes
expensive (for example, if the render error panel might search
for custom views in the project to suggest typo corrections), and
it is currently done under a readlock (because it uses PSI for
some of its error suggestions code).

This CL makes the render error panel perform most of its computation
without holding a readlock, and only grabs it briefly when it needs
to consult PSI. This should help make the interactive performance
while typing partial layout XML feel better.

Change-Id: I14ddb2871f32fd6da494327545de3e6ac2a4629f