Make the resource notification manager smarter
This CL ports some of the extra logic performed in
the layout preview manager's PSI listener to the
resource notification manager. This makes the resource
notification manager skip events known not to affect
resources: edits in comments, whitespace, edits in
attributes that do not have a value node (e.g. in the
process of being typed in by the user) etc.

Change-Id: I8c69dbf61fa68cea8a5e6f8615b9cb726003ee66