Fixing behavior after switching away from preview channel
Previously, if you installed a preview package and then switched back
to the stable channel, updates to the preview would not be shown in
the sdk manager UI, but would be installed if you tried to install
or update something else.
This change causes them to be neither shown nor installed.

Change-Id: I1e89d4f0bf02ca62746f9f597bde3f8f81d9a021