Handle multiple pending full-scans in the resource notification manager
This CL fixes a bug in the ResourceNotificationManager where edits
that trigger a full rescan (such as creating a new resource value via
AndroidResourceUtil, or calling XmlTag.setEscapedText) could result in
only the first edit resulting in a resource notification.

Here's what happens:

(1) An edit arrives from the PSI system with incomplete event
information, so the ResourceFolderRepository schedules a full file
rescan in the event dispatch queue.

(2) The ResourceNotificationManager sees the same event and schedules
a notification update in the event dispatch queue.

(3) An edit arrives from the PSI system for the *second* file. Again,
the ResourceFolderRepository schedules a full file rescan for this in
the event dispatch -- but note that this event is sitting *after* the
ResourceNotificationManager update event in the queue.

(4) The ResourceNotificationManager sees the same event, but because
an update is already pending, it does nothing. This is different from
the ResourceFolderRepository, which tracks rescans on a per file
basis.  The ResourceNotificationManager does not have that fine
granularity; it watches changes across all files in the project.

(5) When the ResourceNotificationManager runs its delayed event, the
second file's rescan has not been processed yet, so it has not yet
rev'ed its version number. The ResourceNotificationManager notices
that it hasn't changed, so it ignores this change.

The fix here is that we need to *always* make sure that the resource
notification manager processes its events after all currently pending
runnables have seen. Therefore, instead of notifying listeners
directly, it enqueues itself a second time in the queue before firing
notification.

This CL also adds a test for this scenario (and a couple others).

Change-Id: Ibd0cc66ff18604d6ed644de57c9ee5adbd0d1a3f