Fix invisible launch emulator radio button getting selected.
When running tests in the google cloud is enabled and the
last device disappears from the list of devices,
the invisible launch emulator radio button would erroneously
be selected. If the user clicks OK in this state, an emulator
will be launched (althouh such option is not even present in
the dialog). This CL fixes this weird behavior.

Change-Id: I498a3b9a38932faab3fe613cfa3a20e08603aa2b