Fixing NPW resizing when moving to second step
When empty, no height is requested by these labels. After the step
is displayed they get populated by <html></html>, which causes them
to request some height. Setting the initial text to <html></html>
so they request the right height from the beginning (before the
dialog is displayed at all).
This seems to partially fix
https://code.google.com/p/android/issues/detail?id=174247
in that the link is moved down somewhat on the text it overlaps.

Change-Id: Ifd23ee8e592845715c97d49c8bf5d5b35da43d03