Fixing concurrency problem during sdk package loading
There was a problem where if, while an sdk refresh is running
and the local load had completed by the remote load had not,
a further non-forcerefresh call came in, that call would return
while only the local results were available instead of waiting
for the complete results.
Changed such that while only the partial results are available
they are always considered out-of-date, and thus we will always
wait until the task has finished.
Change-Id: If0ef086d740b5e990a39d731cddfac0865c0dbc2