Fixing SDK manager flickering
If no user addon sources were present, we weren't updating the
sources map to indicate that there were none. Then when checking
if they had been loaded it looked like they had not, and so we
loaded again, kicking off the change listener again which started
another load...
Eventually the cycle could break, since each load happens on a
different thread, and if the new load started before the previous
was complete it would just add its callbacks to the existing load.
But as long as the first load finished before the second was actually
kicked off, the cycle would continue.

Change-Id: Iecfed70f6cb3df4a3bdb55d30e86c906ecba94d6