Better handling of partially-loaded SDK
https://android-review.googlesource.com/154958 attempted to fix
a problem where sometimes only local sdk packages were available
in the completed callbacks, but was confusing and didn't work all
the time.
In this change we no longer update myPackages in SdkState until the
load has completed, so myPackages always contains only the complete
set.
Client that use the onLocalComplete callback now need to use the
passed-in SdkPackages to retrieve the updated local packages.

Change-Id: Icea87c1f312fe7cc2db52fdaa12c09660e945557