process chooser: Cleanup behavior around previous selection
The process chooser dialog shows a tree with a list of debuggable
processes. Whenever this tree is refreshed, the previous selection
was restored. However, the previous selection was updated only
when the OK button was pressed, so we were always restoring the
previous app that was debugged rather than the previous selection.

This CL saves the current selection every time it changes, and
restores that selection after tree updates.

Fixes https://code.google.com/p/android/issues/detail?id=177232

Change-Id: If6ade213dd217182fce6766f3e1f7d0a2ab9c063