Fixing SDK caching
There were some problems with caching of AndroidSdkData and SdkState.
Previously, when you saved changes in the new sdk manager, it always
set the sdk path, whether it changed or not. This would call a bunch
of code, causing some unnecessarily invalidation/refreshing. One such
effect was that AndroidSdkUtils would get a new (substantially
identical) instance of AndroidSdkData.
When looking up an SdkState, we were doing a reference comparison
on AndroidSdkData, rather than .equals. This meant that, if we use
AndroidSdkUtils.tryToChooseAndroidSdk() to look up an SdkState, we
might get a different SdkState from one call to the next. And since
the remote sdk info is stored in SdkState whereas the local info is
in AndroidSdkData, sometimes inconsistent results were seen. One
effect of this was that sometimes newly installed or deleted packages
weren't correctly reflected in SdkPackages.
This should fix
https://code.google.com/p/android/issues/detail?id=175087.
I didn't reproduce the reported problem exactly, but could get the
same error through a different path.

Change-Id: I860e2471105821b7c9e038ab6534000802a69cd9