Hide the "configure plugins" menu option
Currently configure plugins doesn't work, since currently
there are no plugin-pages configured in
AndroidStudioApplicationInfo.xml. The reason for that is that
StartupWizardModel (an intellij class, invoked by this action)
requires that there be a catch-all plugins page defined, but
there is no way for us to hide the android plugin (which is
obviously required for studio to work). There was previously
support for such a list of required plugins, but it was removed
from the intellij platform by change d06b4bf.
I'm looking into getting it added back.

Change-Id: I9badbff024a2485fbb19d7efe29875a633ccbbc4